Move E2E tests into their own directory, make match one also deal with AI speed to avoid timeout
